refactor: move `types` to the `api` package
