+++ packages/velocity-exchange-verifiers/src/impl/errors/error-builder.ts
@@ -4,7 +4,7 @@
  * Copyright 2022 Velocity Career Labs inc.
  * SPDX-License-Identifier: Apache-2.0
  */
-import { VerificationError } from '../types';
+import { VerificationError } from 'api/types';
 import { ErrorCode } from './error-code';
 
 /**
