+++ packages/velocity-exchange-verifiers/src/impl/utils/index.ts
@@ -4,7 +4,7 @@
  * Copyright 2022 Velocity Career Labs inc.
  * SPDX-License-Identifier: Apache-2.0
  */
-import { VerificationContext } from 'impl/types';
+import { VerificationContext } from 'api/types';
 
 /**
  * Appends a nested path segment to an existing {@link VerificationContext}.
