+++ packages/velocity-exchange-verifiers/src/impl/verifiers/pure-verifiers/alg-is-supported-verifier.ts
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { W3CCredentialJwtV1, Verifier } from 'impl/types';
+import { W3CCredentialJwtV1, Verifier } from 'api/types';
 import { buildError, ERROR_CODES } from 'impl/errors';
 
 /**
