+++ packages/velocity-exchange-verifiers/src/impl/verifiers/pure-verifiers/credential-schema-verifier.ts
@@ -5,7 +5,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { buildError, ERROR_CODES } from 'impl/errors';
-import { W3CCredentialJwtV1, Verifier } from 'impl/types';
+import { W3CCredentialJwtV1, Verifier } from 'api/types';
 
 /**
  * Verifies that the `vc.credentialSchema` field exists in the Credential JWT payload.
