+++ packages/velocity-exchange-verifiers/test/impl/verifiers/pure-verifiers/credential-schema-verifier.test.ts
@@ -7,7 +7,7 @@
 
 import { credentialSchemaVerifier } from 'impl/verifiers/pure-verifiers/credential-schema-verifier';
 import { ERROR_CODES } from 'impl/errors';
-import { W3CCredentialJwtV1, VerificationContext } from 'impl/types';
+import { W3CCredentialJwtV1, VerificationContext } from 'api/types';
 
 describe('credentialSchemaVerifier', () => {
   const baseContext: VerificationContext = {
