+++ packages/velocity-exchange-verifiers/test/impl/verifiers/pure-verifiers/credential-status-verifier.test.ts
@@ -7,7 +7,7 @@
 
 import { credentialStatusVerifier } from 'impl/verifiers/pure-verifiers/credential-status-verifier';
 import { ERROR_CODES } from 'impl/errors';
-import { W3CCredentialJwtV1, VerificationContext } from 'impl/types';
+import { W3CCredentialJwtV1, VerificationContext } from 'api/types';
 
 describe('credentialStatusVerifier', () => {
   const baseContext: VerificationContext = {
