+++ packages/velocity-exchange-verifiers/test/impl/verifiers/pure-verifiers/iss-claim-matches-either-metadata-or-credential-issuer-verifier.test.ts
@@ -7,7 +7,7 @@
 
 import { issClaimMatchesEitherMetadataOrCredentialIssuerVerifier } from 'impl/verifiers/pure-verifiers';
 import { ERROR_CODES } from 'impl/errors';
-import { W3CCredentialJwtV1, VerificationContext } from 'impl/types';
+import { W3CCredentialJwtV1, VerificationContext } from 'api/types';
 
 describe('issClaimMatchesEitherMetadataOrCredentialIssuerVerifier', () => {
   const context: VerificationContext = {
