+++ packages/velocity-exchange-verifiers/test/impl/verifiers/pure-verifiers/iss-claim-matches-metadata-verifier.test.ts
@@ -6,7 +6,7 @@
  */
 
 import { issClaimMatchesMetadataVerifier } from 'impl/verifiers/pure-verifiers';
-import { W3CCredentialJwtV1, VerificationContext } from 'impl/types';
+import { W3CCredentialJwtV1, VerificationContext } from 'api/types';
 import { ERROR_CODES } from 'impl/errors';
 
 describe('issClaimMatchesMetadataVerifier', () => {
